requirements.txt
feedparser>=6.0.0
anthropic>=0.25.0
notion-client>=2.2.1,<3.0.0
tweepy>=4.14.0
python-dotenv>=1.0.0
